Revisió dels topònims dels cadastres dels municipis de Catalunya Nord feta per Joan Becat

Inclou els topònims anteriors i els corregits així com el topònim que apareix al GEOPORTAIL de l'IGN.

Ús

becat_cadastre

Format

El format és una llista de 236 elements, corresponents a la revisió del cadastre de cada municipi. Per cada municipi hi ha un «data frame» amb tantes observacions com topònims i 5 variables:

TOPÒNIMS I FULLS CADASTRALS

Referència del full cadastral del topònim.

NOM*

Aquesta variable no és igual en tots els municipis i pren 3 formes diferents: NOM DEL CADASTRE ANTERIOR I SOBRE EL CADASTRE ACTUAL, NOM SOBRE EL CADASTRE ACTUAL o NOM SOBRE EL CADASTRE ANTERIOR.

NOM*

Aquesta variable no és igual en tots els municipis i pren 3 formes diferents: NOM A CORREGIR SOBRE EL CADASTRE, NOM CORREGIT SOBRE EL CADASTRE ACTUAL, NOM PROPOSAT PEL BATLLE I NO CORREGIT PER LA DGI o NOM SOBRE EL CADASTRE ACTUAL CORREGIT.

NOM SOBRE GÉOPORTAIL (portal cartogràfic de l’IGN)

Topònim que apareix a https://www.geoportail.gouv.fr.

tipus

Tipus de l'element associat al topònim.

Detalls

Els cadastres d'alguns municipis són parcialment corregits i els topònims són dividits en dues taules segons si estan o no corregits (vegeu exemples).

Font

Joan Becat https://joanbecat.cat. Extracció de les dades a data-raw/becat_cadastre.R.

Vegeu també

Exemples

str(becat_cadastre[1])
#> List of 1
#>  $ Aiguatèbia i Talau:'data.frame':	250 obs. of  5 variables:
#>   ..$ TOPÒNIMS I FULLS CADASTRALS                       : chr [1:250] "A4" "200B1" "A5" "200A1" ...
#>   ..$ NOM SOBRE EL CADASTRE ANTERIOR                    : chr [1:250] "Las Artigues" "La Bach" "Bac de Bourdouil Est/Ouest" "Bac de la Pègue" ...
#>   ..$ NOM SOBRE EL CADASTRE ACTUAL CORREGIT             : chr [1:250] "les Artigues" "el Bac" "Bac de Bordoll Est/Ouest" "Bac de la Pega" ...
#>   ..$ NOM SOBRE GÉOPORTAIL (portal cartogràfic de l’IGN): chr [1:250] NA NA "Bac de Bordoll" "Bac de la Pega" ...
#>   ..$ tipus                                             : chr [1:250] "Noms de seccions cadastrals" "Noms de seccions cadastrals" "Noms de seccions cadastrals" "Noms de seccions cadastrals" ...

# Diversitat de capçaleres
unique(lapply(becat_cadastre, names))
#> [[1]]
#> [1] "TOPÒNIMS I FULLS CADASTRALS"                       
#> [2] "NOM SOBRE EL CADASTRE ANTERIOR"                    
#> [3] "NOM SOBRE EL CADASTRE ACTUAL CORREGIT"             
#> [4] "NOM SOBRE GÉOPORTAIL (portal cartogràfic de l’IGN)"
#> [5] "tipus"                                             
#> 
#> [[2]]
#> [1] "TOPÒNIMS I FULLS CADASTRALS"                       
#> [2] "NOM SOBRE EL CADASTRE ACTUAL"                      
#> [3] "NOM A CORREGIR SOBRE EL CADASTRE"                  
#> [4] "NOM SOBRE GÉOPORTAIL (portal cartogràfic de l’IGN)"
#> [5] "tipus"                                             
#> 
#> [[3]]
#> [1] "TOPÒNIMS I FULLS CADASTRALS"                         
#> [2] "NOM DEL CADASTRE ANTERIOR I SOBRE EL CADASTRE ACTUAL"
#> [3] "NOM PROPOSAT PEL BATLLE I NO CORREGIT PER LA DGI"    
#> [4] "NOM SOBRE GÉOPORTAIL (portal cartogràfic de l’IGN)"  
#> [5] "tipus"                                               
#> 
#> [[4]]
#> [1] "TOPÒNIMS I FULLS CADASTRALS"                       
#> [2] "NOM SOBRE EL CADASTRE ANTERIOR"                    
#> [3] "NOM CORREGIT SOBRE EL CADASTRE ACTUAL"             
#> [4] "NOM SOBRE GÉOPORTAIL (portal cartogràfic de l’IGN)"
#> [5] "tipus"                                             
#> 
#> [[5]]
#> [1] "TOPÒNIMS I FULLS CADASTRALS"                       
#> [2] "NOM SOBRE EL CADASTRE ANTERIOR"                    
#> [3] "NOM SOBRE EL CADASTRE ACTUAL NO CORREGIT"          
#> [4] "NOM SOBRE GÉOPORTAIL (portal cartogràfic de l’IGN)"
#> [5] "tipus"                                             
#> 

# Municipis parcialment corregits
grep("_", names(becat_cadastre), value = TRUE)
#>  [1] "Catllà_CADASTRE"                 "Catllà_PROPOSAT"                
#>  [3] "Conat i Vellans_CADASTRE"        "Conat i Vellans_PROPOSAT"       
#>  [5] "Cornellà de Conflent_CADASTRE"   "Cornellà de Conflent_PROPOSAT"  
#>  [7] "els Masos_CADASTRE"              "els Masos_PROPOSAT"             
#>  [9] "Formiguera_CADASTRE"             "Formiguera_PROPOSAT"            
#> [11] "la Roca de l'Albera_CORREGIT"    "la Roca de l'Albera_NO CORREGIT"
#> [13] "Marqueixanes_CADASTRE"           "Marqueixanes_PROPOSAT"          
#> [15] "Mosset_CADASTRE"                 "Mosset_PROPOSAT"                
#> [17] "Orellà_CADASTRE"                 "Orellà_PROPOSAT"                
#> [19] "Rià i Cirac_CADASTRE"            "Rià i Cirac_PROPOSAT"           
#> [21] "Saorra i Toren_CADASTRE"         "Saorra i Toren_PROPOSAT"        
#> [23] "Taurinyà_CADASTRE"               "Taurinyà_PROPOSAT"              
#> [25] "Vallestàvia_CADASTRE"            "Vallestàvia_PROPOSAT"           

# Municipis inclosos
names(becat_cadastre)
#>   [1] "Aiguatèbia i Talau"                   
#>   [2] "Alenyà"                               
#>   [3] "Angostrina i Vilanova de les Escaldes"
#>   [4] "Ansinhan"                             
#>   [5] "Arboçols"                             
#>   [6] "Argelers de la Marenda"               
#>   [7] "Arles"                                
#>   [8] "Bages"                                
#>   [9] "Baixàs"                               
#>  [10] "Banyuls de la Marenda"                
#>  [11] "Banyuls dels Aspres"                  
#>  [12] "Baó"                                  
#>  [13] "Belhestar"                            
#>  [14] "Bolquera"                             
#>  [15] "Bonpàs"                               
#>  [16] "Brullà"                               
#>  [17] "Bula d'Amunt"                         
#>  [18] "Bulaternera"                          
#>  [19] "Cabestany"                            
#>  [20] "Calce"                                
#>  [21] "Calmella"                             
#>  [22] "Cameles"                              
#>  [23] "Campome"                              
#>  [24] "Camporsin"                            
#>  [25] "Canavelles"                           
#>  [26] "Canet"                                
#>  [27] "Cànoes"                               
#>  [28] "Caramanh"                             
#>  [29] "Casafabre"                            
#>  [30] "Cases de Pena"                        
#>  [31] "Cassanhas"                            
#>  [32] "Castell"                              
#>  [33] "Castellnou"                           
#>  [34] "Catllà_CADASTRE"                      
#>  [35] "Catllà_PROPOSAT"                      
#>  [36] "Caudiers de Conflent"                 
#>  [37] "Caudièrs de Fenolhet"                 
#>  [38] "Ceret"                                
#>  [39] "Cervera"                              
#>  [40] "Clairà"                               
#>  [41] "Clarà i Villerac"                     
#>  [42] "Codalet"                              
#>  [43] "Conat i Vellans_CADASTRE"             
#>  [44] "Conat i Vellans_PROPOSAT"             
#>  [45] "Corbera"                              
#>  [46] "Corbera la Cabana"                    
#>  [47] "Cornellà de Conflent_CADASTRE"        
#>  [48] "Cornellà de Conflent_PROPOSAT"        
#>  [49] "Cornellà de la Ribera"                
#>  [50] "Cornellà del Bercol"                  
#>  [51] "Cortsaví"                             
#>  [52] "Cotlliure"                            
#>  [53] "Custoja"                              
#>  [54] "Dorres"                               
#>  [55] "Èguet"                                
#>  [56] "Eina"                                 
#>  [57] "el Barcarès"                          
#>  [58] "el Pertús"                            
#>  [59] "el Soler"                             
#>  [60] "el Tec"                               
#>  [61] "el Voló"                              
#>  [62] "Elna"                                 
#>  [63] "els Angles"                           
#>  [64] "els Banys d'Arles i Palaldà"          
#>  [65] "els Masos_CADASTRE"                   
#>  [66] "els Masos_PROPOSAT"                   
#>  [67] "Enveig"                               
#>  [68] "Er"                                   
#>  [69] "Escaró"                               
#>  [70] "Espirà de Conflent"                   
#>  [71] "Espirà de l'Aglí"                     
#>  [72] "Estagell"                             
#>  [73] "Estavar"                              
#>  [74] "Estoer"                               
#>  [75] "Eus i Coma"                           
#>  [76] "Felhuns"                              
#>  [77] "Fenolhet"                             
#>  [78] "Fillols"                              
#>  [79] "Finestret"                            
#>  [80] "Font-rabiosa"                         
#>  [81] "Font-romeu Odelló Vià"                
#>  [82] "Fontpedrosa"                          
#>  [83] "Formiguera_CADASTRE"                  
#>  [84] "Formiguera_PROPOSAT"                  
#>  [85] "Forques"                              
#>  [86] "Fòssa"                                
#>  [87] "Fullà"                                
#>  [88] "Glorianes"                            
#>  [89] "Illa"                                 
#>  [90] "Jóc"                                  
#>  [91] "Jújols"                               
#>  [92] "l'Albera"                             
#>  [93] "l'Esquerda"                           
#>  [94] "la Bastida"                           
#>  [95] "la Cabanassa"                         
#>  [96] "la Guingueta d'Ix"                    
#>  [97] "la Llaguna"                           
#>  [98] "la Menera"                            
#>  [99] "la Roca de l'Albera_CORREGIT"         
#> [100] "la Roca de l'Albera_NO CORREGIT"      
#> [101] "la Tor de França"                     
#> [102] "la Tor de Querol"                     
#> [103] "la Torre del Bisbe"                   
#> [104] "Lansac"                               
#> [105] "Le Vivièr"                            
#> [106] "les Cluses"                           
#> [107] "Llauró"                               
#> [108] "Llo"                                  
#> [109] "Llupià"                               
#> [110] "Marqueixanes_CADASTRE"                
#> [111] "Marqueixanes_PROPOSAT"                
#> [112] "Matamala"                             
#> [113] "Maurin"                               
#> [114] "Mentet"                               
#> [115] "Millars"                              
#> [116] "Molig"                                
#> [117] "Montalban"                            
#> [118] "Montboló"                             
#> [119] "Montescot"                            
#> [120] "Montesquiu"                           
#> [121] "Montferrer"                           
#> [122] "Montlluís"                            
#> [123] "Montner"                              
#> [124] "Montoriol"                            
#> [125] "Morellàs i les Illes"                 
#> [126] "Mosset_CADASTRE"                      
#> [127] "Mosset_PROPOSAT"                      
#> [128] "Naüja"                                
#> [129] "Nefiac"                               
#> [130] "Noedes"                               
#> [131] "Nyer"                                 
#> [132] "Oceja"                                
#> [133] "Oleta i Èvol"                         
#> [134] "Oms"                                  
#> [135] "Òpol i Perellós"                      
#> [136] "Orbanyà"                              
#> [137] "Orellà_CADASTRE"                      
#> [138] "Orellà_PROPOSAT"                      
#> [139] "Ortafà"                               
#> [140] "Paçà"                                 
#> [141] "Palau de Cerdanya"                    
#> [142] "Palau del Vidre"                      
#> [143] "Paretstortes"                         
#> [144] "Perpinyà"                             
#> [145] "Pesilhan de Conflent"                 
#> [146] "Pesillà de la Ribera"                 
#> [147] "Pi"                                   
#> [148] "Pià"                                  
#> [149] "Planès"                               
#> [150] "Planesas"                             
#> [151] "Pollestres"                           
#> [152] "Pontellà i Nyils"                     
#> [153] "Porta"                                
#> [154] "Portè"                                
#> [155] "Portvendres"                          
#> [156] "Prada"                                
#> [157] "Prats de Molló"                       
#> [158] "Prats de Sornian"                     
#> [159] "Prunet i Bellpuig"                    
#> [160] "Prunhanas"                            
#> [161] "Puigvaledor"                          
#> [162] "Queixàs"                              
#> [163] "Ral"                                  
#> [164] "Ralleu"                               
#> [165] "Rasiguèras"                           
#> [166] "Rebolhet"                             
#> [167] "Reiners"                              
#> [168] "Rià i Cirac_CADASTRE"                 
#> [169] "Rià i Cirac_PROPOSAT"                 
#> [170] "Ribesaltes"                           
#> [171] "Rigardà"                              
#> [172] "Rodés"                                
#> [173] "Salelles"                             
#> [174] "Salses"                               
#> [175] "Sansà"                                
#> [176] "Sant Andreu de Sureda"                
#> [177] "Sant Cebrià"                          
#> [178] "Sant Esteve del Monestir"             
#> [179] "Sant Feliu d'Amunt"                   
#> [180] "Sant Feliu d'Avall"                   
#> [181] "Sant Genís de Fontanes"               
#> [182] "Sant Hipòlit"                         
#> [183] "Sant Joan la Cella"                   
#> [184] "Sant Joan Pla de Corts"               
#> [185] "Sant Llorenç de Cerdans"              
#> [186] "Sant Llorenç de la Salanca"           
#> [187] "Sant Marçal"                          
#> [188] "Sant Martin"                          
#> [189] "Sant Miquel de Llotes"                
#> [190] "Sant Pau de Fenolhet"                 
#> [191] "Sant Pere dels Forcats"               
#> [192] "Sant Nazari"                          
#> [193] "Santa Coloma"                         
#> [194] "Santa Llocaia"                        
#> [195] "Santa Maria la Mar"                   
#> [196] "Saorra i Toren_CADASTRE"              
#> [197] "Saorra i Toren_PROPOSAT"              
#> [198] "Sautó i Fetges"                       
#> [199] "Serdinyà"                             
#> [200] "Serrallonga"                          
#> [201] "Soanyes i Marians"                    
#> [202] "Sornian"                              
#> [203] "Sureda"                               
#> [204] "Talteüll"                             
#> [205] "Tarerac"                              
#> [206] "Targasona"                            
#> [207] "Taurinyà_CADASTRE"                    
#> [208] "Taurinyà_PROPOSAT"                    
#> [209] "Tellet"                               
#> [210] "Terrats"                              
#> [211] "Tesà"                                 
#> [212] "Teulís"                               
#> [213] "Toès i Entrevalls"                    
#> [214] "Toluges"                              
#> [215] "Torderes"                             
#> [216] "Torrelles"                            
#> [217] "Trasserra"                            
#> [218] "Trevilhac"                            
#> [219] "Trilhan"                              
#> [220] "Trullars"                             
#> [221] "Tuïr"                                 
#> [222] "Ur"                                   
#> [223] "Vallcebollera"                        
#> [224] "Vallestàvia_CADASTRE"                 
#> [225] "Vallestàvia_PROPOSAT"                 
#> [226] "Vallmanya"                            
#> [227] "Vernet"                               
#> [228] "Vilallonga de la Salanca"             
#> [229] "Vilallonga dels Monts"                
#> [230] "Vilamulaca"                           
#> [231] "Vilanova de la Ribera"                
#> [232] "Vilanova de Raó"                      
#> [233] "Vinçà"                                
#> [234] "Vingrau"                              
#> [235] "Viran"                                
#> [236] "Vivers"