Descripció dels fitxers pdf originals

Informació extreta de l'apartat de la web amb revisions dels topònims dels cadastres dels municipis de Catalunya Nord.

Ús

fitxers_cadastre

Format

Un «data frame» amb 224 observacions de les següents 4 variables.

municipi

Nom del municipi.

url

URL de la pàgina del municipi.

url_pdf

URL del fitxer en format pdf de la revisió de topònims del municipi.

md5sum

Valor dels resums del fitxer pdf. Útil per detectar actualitzacions dels fitxers.

Detalls

L'enllaç del pdf de Vilafranca de Conflent apunta a la url del poble i no hi ha pdf disponible. La url del pdf de la web de Corbera la Cabanal és incorrecte (li falta la `f` final). No hi ha disponibles en català els pdf de Calce, Espirà de l’Aglí i Jújols i s'usa la versió en francés.

Font

Joan Becat https://joanbecat.cat. Extracció de les dades a data-raw/fitxers_cadastre.R.

Vegeu també

Exemples

str(fitxers_cadastre)
#> 'data.frame':	224 obs. of  4 variables:
#>  $ municipi: chr  "Aiguatèbia" "Alenyà" "Angostrina i Vilanova de les Escaldes" "Ansinhan" ...
#>  $ url     : chr  "https://joanbecat.cat/ca/aiguatebia/" "https://joanbecat.cat/ca/alenya-2/" "https://joanbecat.cat/ca/angostrina-i-vilanova-de-les-escaldes/" "https://joanbecat.cat/ca/ansinhan/" ...
#>  $ url_pdf : chr  "https://joanbecat.cat/wp-content/uploads/2019/07/cat-Web-jbk-cadastre_Aiguate%CC%80bia.pdf" "https://joanbecat.cat/wp-content/uploads/2019/07/cat-Web-jbk-cadastre_Alenya%CC%80.pdf" "https://joanbecat.cat/wp-content/uploads/2019/08/cat-Web-jbk-cadastre_Angostrina.pdf" "https://joanbecat.cat/wp-content/uploads/2019/08/cat-Web-jbk-cadastre_Ancinhan_noCorregit-4.pdf" ...
#>  $ md5sum  : chr  "921d57d54c52ae303497c6bb23482149" "4d31d748b181dd7c6a18428450545acf" "6c01e52adabd3de866a5cdc35174b326" "3601663ef245dfaefc380000d2c1f7a4" ...