Introduccions de les revisions dels topònims dels cadastres dels municipis de Catalunya Nord feta per Joan Becat

Contingut de la introducció dels fitxers en prosa.

Ús

intro_cadastre

Format

El format és una llista de 223 elements, corresponents a la revisió del cadastre de cada municipi. Per cada municipi hi ha un vector de caràcters amb el contingut de la introducció del fitxers.

Detalls

Les introduccions contenen informació sobre l'estat del cadastre i de GEOPORTAIL (revisat o no), data en què es va fer la revisió, participants en la revisió i la referència a l'«Atles toponímic de Catalunya Nord - Atlas toponymique de Catalogne Nord, ed. Terra Nostra».

Font

Joan Becat https://joanbecat.cat. Extracció de les dades a data-raw/becat_cadastre.R.

Exemples

str(intro_cadastre)
#> List of 223
#>  $ Aiguatèbia                           : chr [1:7] "Cadastre corregit el 28 de juny del 2017 a la Casa de la Vila. Batlle: Georges Vicens." "Comanda del Consell Departamental del Pirineu Oriental. Comissió municipal: Lucien MITJAVILLE, antic batlle, Ge"| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Alenyà                               : chr [1:4] "Cadastre corregit el 19 de setembre del 1989 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà. Batlle: "| __truncated__ "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia d’Alenyà sobre el seu portal internet GEOPO"| __truncated__ "Nota: Pel conjunt dels topònims d’Alenyà veure l’Atles toponímic de Catalunya Nord - Atlas toponymique de Catal"| __truncated__
#>  $ Angostrina i Vilanova de les Escaldes: chr [1:6] "Cadastre corregit el 21 de juny del 2017 a la Casa de la Vila amb Jacques BARDES. Batllessa: Hélène JOSENDE. Co"| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "Recull preliminar dels noms del municipi el 1983 amb Marcel DOUTRES, i els 26 i 27 d’agost del 2010 amb Jacques"| __truncated__ ...
#>  $ Ansinhan                             : chr [1:2] "El cadastre del municipi d’Ansinhan no és corregit." "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia d’Ansinhan sobre el seu mapa a 1:25.000 "| __truncated__
#>  $ Arboçols                             : chr [1:7] "Cadastre corregit el 18 de juliol del 2017 a la Casa de la Vila. Batlle: Étienne Surjus. Comissió municipal: Gi"| __truncated__ "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Argelers de la Marenda               : chr [1:3] "El cadastre del municipi d’Argelers de la Marenda no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia d’Argelers de la Marenda sobre el seu porta"| __truncated__ "Nota: Trobareu el conjunt dels topònims d’Argelers de la Marenda amb la seua grafia catalana correcta, dins l’A"| __truncated__
#>  $ Arles                                : chr [1:6] "Cadastre corregit el 9 de novembre del 2018 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà. Batlle: R"| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "Primera correcció dels noms del cadastre al novembre del 2012. Batlle: André ALA. Comissió municipal: Francesc "| __truncated__ ...
#>  $ Bages                                : chr [1:7] "Cadastre corregit el 20 de desembre del 1990 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà. Batlle: "| __truncated__ "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "Cadastre corregit el juliol del 2012 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà. Batlle: Serge SO"| __truncated__ "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Baixàs                               : chr [1:5] "Cadastre corregit el 9 d’octubre del 2006 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Gilles FOXONET. Comissió municipal: Martine BILLÈS, Charles BOBO, Marie DUMAS, adjunta, Jacqueline OUIL"| __truncated__ "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Baixàs sobre el seu portal internet GEOP"| __truncated__ ...
#>  $ Banyuls de la Merenda                : chr [1:3] "El cadastre del municipi de Banyuls de la Marenda no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Banyuls de la Marenda sobre el seu porta"| __truncated__ "Nota: Trobareu el conjunt dels topònims de Banyuls de la Marenda, amb la seua grafia catalana correcta, dins l’"| __truncated__
#>  $ Banyuls dels Aspres                  : chr [1:3] "El cadastre del municipi de Banyuls dels Aspres no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Banyuls dels Aspres sobre el seu portal "| __truncated__ "Nota: Trobareu el conjunt dels topònims de Banyuls des Aspres amb la seua grafia catalana correcta, dins l’Atle"| __truncated__
#>  $ Baó                                  : chr [1:6] "Cadastre corregit al desembre del 1983 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Guy CASADEVALL. Comissió municipal: Guy CASADEVALL, batlle, Roger CAZENOVE, Michel DANJOU, Adolphe JOUÉ"| __truncated__ "Brigada regional de topografia: Antoine BORDIGNON, enginyer del cadastre. DGI de Perpinyà: Michel PONTIÉ." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Belhestar                            : chr [1:2] "El cadastre del municipi de Belhestar no és corregit." "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia de Belhestar sobre el seu mapa a 1:25.00"| __truncated__
#>  $ Bolquera                             : chr [1:5] "Cadastre corregit el 15 de juny del 2011 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Jean-Pierre ABEL. Comissió municipal: Denise BASSO, Jean-Louis BRUNET, adjunt, Jackie COLL, adjunt, Gil"| __truncated__ "Recull preliminar dels noms del municipi el 5 d’octubre del 2004, a la Casa de la Vila, amb Monique HERNANDEZ, "| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Bolquera sobre el seu portal internet GE"| __truncated__ ...
#>  $ Bonpàs                               : chr [1:5] "Cadastre corregit el 26 de març del 1996 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Jean-Paul BATLLE. Comissió municipal: Louis PUJOL, batlle adjunt, Sylvie ALBERT, Geneviève BOURRET, Jea"| __truncated__ "Brigada regional de topografia: Patrick MOUREY, enginyer del cadastre. DGI de Perpinyà: Michel PONTIÉ. Expert: "| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Bonpàs sobre el seu portal internet GEOP"| __truncated__ ...
#>  $ Brullà                               : chr [1:3] "El cadastre del municipi de Brullà no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Brullà sobre el seu portal internet GEOP"| __truncated__ "Nota: Trobareu el conjunt dels topònims de Brullà, amb la seua grafia catalana correcta, dins l’Atles toponímic"| __truncated__
#>  $ Bula d’Amunt                         : chr [1:7] "Cadastre corregit el 6 de setembre del 2017 a la Casa de la Vila amb Xavier MENDOZA. Batlle: Yann OHEIX." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Bulaternera                          : chr [1:7] "Cadastre corregit el 22 de juny del 2017 a la Casa de la Vila. Batlle: Jean PAYROU." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Cabestany                            : chr [1:5] "Cadastre corregit el 26 de març del 1985 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Jean VILA. Comissió municipal: Andreu BAUX, adjunt, Ramon DARDENNE, Víctor ORTIZ Brigada regional de to"| __truncated__ "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Cabestany sobre el seu portal internet G"| __truncated__ ...
#>  $ Calce                                : chr [1:3] "Le cadastre de la commune de Calce n’est pas corrigé." "L’Institut Géographique National (IGN, Paris) a corrigé la toponymie de Calce sur son portail internet GEOPORTA"| __truncated__ "Note: Vous trouverez l’ensemble des toponymes de Calce avec leur graphie catalane correcte dans l’Atles toponím"| __truncated__
#>  $ Calmella                             : chr [1:4] "El cadastre del municipi de Calmella no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Calmella sobre el seu mapa a 1:25.000 Ce"| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Calmella sobre el seu portal internet GE"| __truncated__ "Nota: Trobareu el conjunt dels topònims de Calmella amb la seua grafia catalana correcta, dins l’Atles toponími"| __truncated__
#>  $ Cameles                              : chr [1:3] "El cadastre del municipi de Cameles no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Cameles sobre el seu portal internet GEO"| __truncated__ "Nota: Trobareu el conjunt dels topònims de Cameles amb la seua grafia catalana correcta, dins l’Atles toponímic"| __truncated__
#>  $ Campome                              : chr [1:6] "Cadastre corregit el 9 de setembre del 2017 a la Casa de la Vila. Batlle: Christophe CAROL." "Comissió municipal: Christophe CAROL, batlle, Jean-Louis BOSC, adjunt, Thèrèse CARON, adjunta, Jacques COMBAUT, adjunt." "Comanda de la Comunitat de Municipis del Conflent-Canigó. Responsable a la DGI de Perpinyà: Patrick MOUREY, eng"| __truncated__ "Recull preliminar dels noms del municipi: 27 de febrer del 2004: Aimé BAGO, adjunt, Jacques COMBAUT, adjunt, Ma"| __truncated__ ...
#>  $ Camporsin                            : chr [1:4] "Cadastre corregit el 17 de febrer del 2017 a la Casa de la Vila. Batlle: Alain BOYER. Enquesta prop de Jean GAL"| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà. Assessor lingüístic: Florian VERNET, APRENE i Un"| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia de Camporsin sobre el seu portal interne"| __truncated__
#>  $ Canavelles                           : chr [1:7] "Cadastre corregit el 10 de juny del 2016 a la Casa de la Vila amb Michel BLEUNVEN, adjunt." "Batlle: Sebastien NENS. Comanda de la Comunitat de municipis Conflent-Canigó." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Canet                                : chr [1:3] "El cadastre del municipi de Canet no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Canet sobre el seu portal internet GEOPO"| __truncated__ "Nota: Trobareu el conjunt dels topònims de Canet, amb la seua grafia catalana correcta, dins l’Atles toponímic "| __truncated__
#>  $ Cànoes                               : chr [1:5] "Cadastre corregit el 26 de març del 1989 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle:Romà ESCUDIÉ. Comissió municipal: Romà ESCUDIÉ, batlle, Miquel GUIMEZANES, Francesc NOELL, Gervasi SOUBI"| __truncated__ "Brigada regional de topografia: Antoine BORDIGNON, enginyer del cadastre. DGI de Perpinyà: Michel PONTIÉ. Exper"| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Cànoes sobre el seu portal internet GEOP"| __truncated__ ...
#>  $ Caramanh                             : chr [1:2] "El cadastre del municipi de Caramanh no és corregit." "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia de Caramany sobre el seu portal internet"| __truncated__
#>  $ Casafabre                            : chr [1:7] "Cadastre corregit el 19 de setembre del 2017 a la Casa de la Vila. Batlle: Claude GOMEZ." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Cases de Pena                        : chr [1:3] "El cadastre del municipi de Cases de Pena no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Cases de Pena sobre el seu portal intern"| __truncated__ "Nota: Trobareu el conjunt dels topònims de Cases de Pena, amb la seua grafia catalana correcta, dins l’Atles to"| __truncated__
#>  $ Cassanhas                            : chr [1:2] "El cadastre del municipi de Cassanhas no és corregit." "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia de Cassanhas sobre el seu portal interne"| __truncated__
#>  $ Castell                              : chr [1:7] "Cadastre corregit el 14 de juny del 2016 a la Casa de la Vila. Batlle: Juliette CASES. Comissió municipal: Juli"| __truncated__ "Comanda de la Comunitat de Municipis de Conflent-Canigó.." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Castellnou                           : chr [1:3] "El cadastre del municipi de Castellnou no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Castellnou sobre el seu portal internet "| __truncated__ "Nota: Trobareu el conjunt dels topònims de Castellnou, amb la seua grafia catalana correcta, dins l’Atles topon"| __truncated__
#>  $ Catllà                               : chr [1:7] "Cadastre corregit el 6 de febrer del 2018 a la Casa de la Vila. Batllessa: Josette PUJOL. Comissió municipal: M"| __truncated__ "Comanda de la Comunitat de Municipis de Conflent-Canigó." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Caudiers de Conflent                 : chr [1:7] "Cadastre corregit el 26 de juny del 2017 a la Casa de la Vila. Batlle: Frédéric BÈS." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Caudièrs de Fenolhet                 : chr [1:2] "El cadastre del municipi de Caudièrs de Fenolhet no és corregit." "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia de Caudièrs de Fenolhet sobre el seu por"| __truncated__
#>  $ Ceret                                : chr [1:5] "Cadastre corregit el 29 d’abril del 1993 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà. Batlle: Henr"| __truncated__ "Brigada regional de topografia: Patrick MOUREY, enginyer del cadastre, Gérard MARTY, geòmetre. DGI de Perpinyà: Louis CABANES." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Ceret sobre el seu portal internet GEOPO"| __truncated__ ...
#>  $ Cervera                              : chr [1:3] "El cadastre del municipi de Cervera no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Cervera sobre el seu portal internet GEO"| __truncated__ "Nota: Trobareu el conjunt dels topònims de Cervera amb la seua grafia catalana correcta, dins l’Atles toponímic"| __truncated__
#>  $ Clairà                               : chr [1:6] "Cadastre corregit el 10 de novembre del 2010 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà. Batlle: "| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "Recull preliminar dels noms del municipi el 1988 amb Marcelle PELLICER, i el 2006, a la Casa de la Vila, amb Ma"| __truncated__ ...
#>  $ Clarà i Villerac                     : chr [1:7] "Cadastre corregit el 12 de juny del 2015 a la Casa de la Vila. Batlle: Robert LAGUERRE." "Comanda de la Comunitat de Municipis de Conflent-Canigó.." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Codalet                              : chr [1:7] "Cadastre corregit el 19 d’octubre del 2009 a iniciativa de la DGI de Perpinyà." "Batlle: Serge JUANCHICH." "DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Conat i Vellans                      : chr [1:7] "Cadastre corregit el 23 de juny del 2016 a la Casa de la Vila. Batlle: Pascal Espeut." "Comanda de la Comunitat de Municipis de Conflent-Canigó." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Corbera la Cabana                    : chr [1:6] "Cadastre corregit el 21 de gener del 2015 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà, amb Francis TORRÈS." "DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "Recull preliminar dels noms del municipi: 17 d’abril del 2010: Jean DALMER, Francis TORRÈS, Joan BECAT." ...
#>  $ Corbera                              : chr [1:7] "Cadastre corregit el 17 de maig del 2017 a la Casa de la Vila. Batlle: Joseph SYLVESTRE. Comissió municipal: Je"| __truncated__ "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Cornellà de Conflent                 : chr [1:7] "Cadastre corregit el 31 de maig del 2016 a la Casa de la Vila. Batlle: Patrice ARRO. Comissió municipal: Patric"| __truncated__ "Comanda de la Comunitat de Municipis de Conflent-Canigó." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Cornellà de la Ribera                : chr [1:8] "Cadastre corregit el 28 d’abril del 2017 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batllessa: Gislène BELTRAN-CHARRE. Comissió municipal: Gislène BELTRAN-CHARRE, batllessa, Marc MADINE, primer adjunt." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." ...
#>  $ Cornellà del Bercol                  : chr [1:4] "Cadastre corregit el 6 de juny del 1990 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Marcel AMOUROUX. Comissió municipal: Robert ALIBERT, Jean CRIBAILLET, Lucien HUC. Brigada regional de t"| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Cornellà del Bercol sobre el seu portal "| __truncated__ "Nota: Trobareu el conjunt dels topònims de Cornellà del Bercol, amb la seua grafia catalana correcta, dins l’At"| __truncated__
#>  $ Cortsaví                             : chr [1:7] "Cadastre corregit el 26 de setembre del 2017 a la Casa de la Vila. Batlle: Antoine CHRYSOSTOME. Comissió munici"| __truncated__ "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Cotlliure                            : chr [1:5] "Cadastre corregit el 10 de desembre del 1992 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà. Batlle: "| __truncated__ "Brigada regional de topografia: Patrick MOUREY, enginyer del cadastre. DGI de Perpinyà: Michel PONTIÉ. Expert: "| __truncated__ "Recull preliminar dels noms del municipi: el 1988: Francis. ORTEGA, Maria Dolors SOLÀ.OMS, Joan Lluís VALLS, Mi"| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Cotlliure sobre el seu portal internet G"| __truncated__ ...
#>  $ Custoja                              : chr [1:5] "Cadastre corregit el 30 d’agost del 2017 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Michel ANRIGO. Comissió municipal: Michel ANRIGO, batlle, Jean-Louis CASANOVA, primer adjunt. Comanda d"| __truncated__ "Recull preliminar dels noms del municipi: 23 d’abril del 2010, a la Casa de la Vila: Jean-Marie MALIGNON, batll"| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Custoja sobre el seu portal internet GEO"| __truncated__ ...
#>  $ Dorres                               : chr [1:9] "Cadastre corregit el 3 de juliol del 2017 a la Casa de la Vila." "Batllessa: Sylvie CANDAU." "Comissió municipal: Sylvie CANDAU, batlle, Victor MARTY, antic batlle." "Comanda del Consell Departamental del Pirineu Oriental." ...
#>  $ Èguet                                : chr [1:3] "El cadastre del municipi d’Èguet no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia d’Èguet sobre el seu portal internet GEOPOR"| __truncated__ "Nota: Trobareu el conjunt dels topònims d’Èguet, amb la seua grafia catalana correcta, dins l’Atles toponímic d"| __truncated__
#>  $ Eina                                 : chr [1:3] "El cadastre del municipi d’Eina no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia d’Eina sobre el seu portal internet GEOPORT"| __truncated__ "Nota: Trobareu el conjunt dels topònims d’Eina, amb la seua grafia catalana correcta, dins l’Atles toponímic de"| __truncated__
#>  $ el Barcarès                          : chr [1:3] "El cadastre del municipi del Barcarès no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia del Barcarès sobre el seu portal internet G"| __truncated__ "Nota: Trobareu el conjunt dels topònims del Barcarès, amb la seua grafia catalana correcta, dins l’Atles toponí"| __truncated__
#>  $ el Pertús                            : chr [1:4] "Cadastre corregit el 16 d’octubre del 2009, a iniciativa de la DGI de Perpinyà, amb Michael PAYROT, conseller. "| __truncated__ "Recull preliminar dels noms del municipi el 9 de desembre del 1999: Joan TOCABENS, Joan BECAT." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia del Pertús sobre el seu portal internet GEO"| __truncated__ "Nota: Trobareu el conjunt dels topònims del Pertús, amb la seua grafia catalana correcta, dins l’Atles toponími"| __truncated__
#>  $ el Soler                             : chr [1:5] "Cadastre corregit el 16 de març del 1992 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: André DAUGNAC. Comissió municipal: Celeste AUJALEU, Gilbert BERNADACH, Adrià BLANIC, Renat COUDOUGNAN, "| __truncated__ "Brigada regional de topografia: Antoine BORDIGNON, enginyer del cadastre. DGI de Perpinyà: Louis CABANES. Exper"| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia del Soler sobre el seu portal internet GEOP"| __truncated__ ...
#>  $ el Tec                               : chr [1:8] "Cadastre corregit el 7 de juliol del 2015 a la Casa de la Vila. Batlle: Guillaume CERVANTES." "Comissió municipal: Guillaume CERVANTES, batlle, Jean-Pierre CASSÉ, adjunt." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." ...
#>  $ el Voló                              : chr [1:6] "Cadastre corregit el 31 de gener del 2003 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Jean-Pierre SALGAS. Comissió municipal: Marc PEUS, adjunt, Alain SANTRAILLE, adjunt, Pierre BÈS, Jean P"| __truncated__ "DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Elna                                 : chr [1:5] "Cadastre corregit el 20 de juny del 1990 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Narcisse PLANAS. Comissió municipal: Gilbert AUGUET, August OMS." "Brigada regional de topografia: Antoine BORDIGNON, enginyer del cadastre. DGI de Perpinyà: Michel PONTIÉ. Exper"| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia d’Elna sobre el seu portal internet GEOPORT"| __truncated__ ...
#>  $ els Angles                           : chr [1:3] "El cadastre del municipi dels Angles no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia dels Angles sobre el seu mapa a 1:25.000 Fo"| __truncated__ "Nota: Trobareu el conjunt dels topònims dels Angles amb la seua grafia catalana correcta, dins l’Atles toponími"| __truncated__
#>  $ els Banys d’Arles i Palaldà          : chr [1:3] "El cadastre del municipi dels Banys d’Arles i Palaldà no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia dels Banys d’Arles i Palaldà sobre el seu p"| __truncated__ "Nota: Trobareu el conjunt dels topònims dels Banys d’Arles i Palaldà, amb la seua grafia catalana correcta, din"| __truncated__
#>  $ els Masos                            : chr [1:7] "Cadastre corregit el 30 de gener del 2018 a la Casa de la Vila. Batlle: Guy CASSOLY." "Comanda de la Comunitat de Municipis de Conflent-Canigó.." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Enveig                               : chr [1:3] "El cadastre del municipi d’Enveig no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia d’Enveig sobre el seu portal internet GEOPO"| __truncated__ "Nota: Trobareu el conjunt dels topònims d’Enveig amb la seua grafia catalana correcta, dins l’Atles toponímic d"| __truncated__
#>  $ Er                                   : chr [1:3] "El cadastre del municipi d’Er no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia d’Er sobre el seu portal internet GEOPORTAI"| __truncated__ "Nota: Trobareu el conjunt dels topònims d’Er, amb la seua grafia catalana correcta, dins l’Atles toponímic de C"| __truncated__
#>  $ Escaró                               : chr [1:7] "Cadastre corregit el 7 de juny del 2016 a la Casa de la Vila. Batlle: André AMBRIGOT." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Espira de Conflent                   : chr [1:8] "Cadastre corregit el 6 de juliol del 2016 a la Casa de la Vila." "Batlle: Roger PAILLÈS." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." ...
#>  $ Espirà de l’Aglí                     : chr [1:5] "Cadastre corrigé le 21 d’abril del 2006 à la mairie, à l’initiative de la DGI de Perpignan. Maire: Gérard BILE."| __truncated__ "DGI de Perpignan: Patrick MOUREY, ingénieur du cadastre." "Expert: Joan BECAT, Études Catalanes, Université de Perpignan." "L’Institut Géographique National (IGN, Paris) a corrigé la toponymie de Espira-de-l’Agly sur son portail intern"| __truncated__ ...
#>  $ Estagell                             : chr [1:7] "Cadastre corregit el 31 de juny del 2015 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle Roger FERRER. Comissió municipal: Roger FERRER, batlle, Joël HENIN, Michel FORNER, Antoine MORCILLO." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Estavar                              : chr [1:3] "El cadastre del municipi d’Estavar no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia d’Estavar sobre el seu portal internet GEOP"| __truncated__ "Trobareu el conjunt dels topònims d’Estavar, amb la seua grafia catalana correcta, dins l’Atles toponímic de Ca"| __truncated__
#>  $ Estoer                               : chr [1:7] "Cadastre corregit el 21 de juliol del 2017 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Louis QUÊS." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Eus i Coma                           : chr [1:6] "Cadastre corregit a iniciativa de la DGI de Perpinyà el 19 de juliol del 2012 a la Casa de la Vila, amb Philipp"| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "Recull preliminar dels noms del municipi l’any 2007 amb Gilbert BOSOM, Philippe DORANDEU i Henri IRLA, i l’any "| __truncated__ ...
#>  $ Felhuns                              : chr [1:4] "Cadastre corregit el 18 de maig del 2016 a la Casa de la Vila. Batlle: Claude FILLOL." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà. Assessor lingüístic: Florian VERNET, APRENE i Un"| __truncated__
#>  $ Fenolhet                             : chr [1:2] "El cadastre del municipi de Fenolhet no és corregit." "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia de Fenolhet sobre el seu portal internet"| __truncated__
#>  $ Fillols                              : chr [1:5] "Cadastre corregit el 23 de juny del 2016 a la Casa de la Vila. Batlle: Claude ESCAPE. Comissió municipal: Julie"| __truncated__ "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "Recull preliminar dels noms del municipi: 6 de juliol del 2000: Bertin CALVET, René TAURINYA, Ramon GUAL, Joan BECAT." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Fillols sobre el seu portal internet GEO"| __truncated__ ...
#>  $ Finestret                            : chr [1:7] "Cadastre corregit el 21 de juny del 2016 a la Casa de la Vila. Batlle: Jean-Michel PAULO." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Font-rabiosa                         : chr [1:8] "Cadastre corregit el 20 de desembre del 2016 a la Casa de la Vila. Batlle: Pierre BATAILLE." "Comissió municipal: Cédric CASTELLO, Jean CRÉGO." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." ...
#>  $ Font-romeu                           : chr [1:5] "Cadastre corregit el 8 d’abril del 1998 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Pierre BOSSELUT. Comissió municipal: Claude CAILLOT, Louis COSTA, Paul GILLES, Albert LAPERC, Jean-François TRICAUD." "Brigada regional de topografia: Patrick MOUREY, enginyer del cadastre. Servei del cadastre de Perpinyà: Michel "| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de la Cabanassa sobre el seu portal interne"| __truncated__ ...
#>  $ Fontpedrosa                          : chr [1:6] "Cadastre corregit l’1 de juny del 2016 a la Casa de la Vila. Batllessa: Arlette BIGORRE.Comissió municipal: Arl"| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "Recull preliminar dels noms del municipi el 5 de març del 2004 amb Maurice CALVET, Albert MARGAIL, i a la Casa "| __truncated__ ...
#>  $ Formiguera                           : chr [1:6] "Cadastre corregit a iniciativa de la DGI de Perpinyà el 19 de novembre del 2013 a la Casa de la Vila, amb Franç"| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "Recull preliminar dels noms del municipi el 7 de novembre del 2006, a la Casa de la Vila, amb Yves BASO, batlle"| __truncated__ ...
#>  $ Forques                              : chr [1:5] "Cadastre corregit el 26 de juny del 1999 a la Casa de la Vila, a iniciativa del municipi. Batlle: Maurice PEYTA"| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de Ceret sobre el seu portal internet GEOPO"| __truncated__ ...
#>  $ Fòssa                                : chr [1:2] "El cadastre del municipi de Fòssa no és corregit." "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia de Fòssa sobre el seu portal internet GE"| __truncated__
#>  $ Fulla                                : chr [1:7] "Cadastre corregit el 3 de juny del 2016 a la Casa de la Vila. Batlle: Pierre BAZÉLY." "Comanda la Comunitat de Municipis del Conflent-Canigó." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Glorianes                            : chr [1:7] "Cadastre corregit el 14 de desembre del 2016 a la Casa de la Vila Batllessa: Céline DRAGUÉ- PAZICAN. Comissió m"| __truncated__ "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Illa                                 : chr [1:7] "Cadastre corregit el 10 de juny del 2000 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Henri SOLER. Comissió municipal: Louis BOURRET, adjunt, Thérèse CABANAT, adjunta, Francis MESTRES, René"| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Jóc                                  : chr [1:7] "Cadastre corregit el 10 de juny del 2016 a la Casa de la Vila. Batlle: Jean-Pierre VILLELONGUE." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ Jújols                               : chr [1:7] "Cadastre corrigé le 22 octobre 2015 à la mairie. Maire: Eric NIVET." "Commande de la Communauté de Communes du Conflent-Canigó." "Responsable à la DGI de Perpignan: Patrick MOUREY, ingénieur du cadastre." "Expert: Joan BECAT, Études Catalanes, Université de Perpignan." ...
#>  $ l’Albera                             : chr [1:6] "Cadastre corregit el 26 de juliol del 2017 a la Casa de la Vila. Batlle: Marc DE BESOMBES- SINGLA. Comanda del "| __truncated__ "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." "Recull preliminar dels noms del municipi: 8 de desembre del 1999: Joan Tocabens, Joan BECAT." ...
#>  $ l’Esquerda                           : chr [1:2] "El cadastre del municipi de l’Esquerda no és corregit." "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia de l’Esquerda sobre el seu portal intern"| __truncated__
#>  $ la Bastida                           : chr [1:7] "Cadastre corregit el 4 de juliol del 2017 a la Casa de la Vila. Batlle: Daniel BAUX." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ la Cabanassa                         : chr [1:3] "El cadastre del municipi de la Cabanassa no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de la Cabanassa sobre el seu portal interne"| __truncated__ "Nota: Trobareu el conjunt dels topònims de la Cabanassa, amb la seua grafia catalana correcta, dins l’Atles top"| __truncated__
#>  $ la Guingueta d’Ix                    : chr [1:5] "Cadastre corregit el 15 de juny del 2011 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Jean-Jacques FORTUNY. Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre. Expert: "| __truncated__ "Recull preliminar dels noms del municipi, a la Casa de la Vila, el 27 d’agost del 2010: Augustin MEYA, adjunt, Joan BECAT." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de la Guingueta d’Ix sobre el seu portal in"| __truncated__ ...
#>  $ la Llaguna                           : chr [1:7] "Cadastre corregit el 7 d’octubre del 2016 a la Casa de la Vila, amb Alain FERRÉ. Batlle: Jean- Pierre ASTRUCH." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ la Menera                            : chr [1:7] "Cadastre corregit el 9 d’octubre del 2016 a la Casa de la Vila. Batllessa: Agnès PARAYRE." "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ la Roca de l’Albera                  : chr [1:8] "Cadastre corregit el 31 de març del 1995 a a iniciativa de la DGI de Perpinyà." "Batllessa: Maryse ARMADA." "DGI de Perpinyà: Michel PONTIÉ." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà." ...
#>  $ la Tor de França                     : chr [1:2] "El cadastre del municipi de la Tor de França no és corregit." "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia de la Tor de França sobre el seu portal "| __truncated__
#>  $ la Tor de Querol                     : chr [1:3] "El cadastre del municipi de la Tor de Querol no és corregit." "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de la Tor de Querol sobre el seu portal int"| __truncated__ "Nota: Trobareu el conjunt dels topònims de la Tor de Querol, amb la seua grafia catalana correcta, dins l’Atles"| __truncated__
#>  $ la Torre del Bisbe                   : chr [1:4] "Cadastre corregit el 26 de març del 1990 a la Casa de la Vila, a iniciativa de la DGI de Perpinyà." "Batlle: Joan Claudi DURAN. Comissió municipal: Josep BERTRAND, Joan Claudi DURAN, batlle, Marcel·lí LAFITTE, Pe"| __truncated__ "L’Institut Geogràfic Nacional (IGN, París) ha corregit la toponímia de la Torre del Bisbe sobre el seu portal i"| __truncated__ "Nota: Trobareu el conjunt dels topònims de la Torre del Bisbe, amb la seua grafia catalana correcta, dins l’Atl"| __truncated__
#>  $ Lansac                               : chr [1:2] "El cadastre del municipi de Lansac no és corregit." "L’Institut Geogràfic Nacional (IGN, París) no ha corregit la toponímia de Lansac sobre el seu portal internet G"| __truncated__
#>  $ le Vivièr                            : chr [1:6] "Cadastre corregit el 15 de juliol del 2015 a la Casa de la Vila. Batlle: Michel BÉNET. Comissió municipal: Mich"| __truncated__ "Comanda del Consell Departamental del Pirineu Oriental." "Responsable a la DGI de Perpinyà: Patrick MOUREY, enginyer del cadastre." "Expert: Joan BECAT, Estudis Catalans, Universitat de Perpinyà. Assessor lingüístic: Florian VERNET, APRENE i Un"| __truncated__ ...
#>   [list output truncated]