Metadades de les revisions dels cadastres

Informació sistematitzada extreta de intro_cadastre.

Ús

meta_cadastre

Format

A data frame with 223 observations on the following 6 variables.

municipi

Nom del municipi.

corregit

Vector logic que indica si el cadastre ha estat corregit.

atles_CatNord

Vector logic que indica si els topònims apareixen a l’Atles toponímic de Catalunya Nord, ed. Terra Nostra, Prada, 2015, 2 volums, 970p.

ign_geoportail

Vector logic que indica si els topònims han estat corregits al GEOPORTAIL de l'IGN.

ign_mapes

a character vector

catala_pdf

Vector logic que indica si el fitxer pdf és en català (TRUE) o en francès (FALSE).

Detalls

Font

Joan Becat https://joanbecat.cat. Extracció de les dades a data-raw/metadades_becat_cadastre.R.

Vegeu també

Exemples

str(meta_cadastre)
#> 'data.frame':	223 obs. of  6 variables:
#>  $ municipi      : chr  "Aiguatèbia" "Alenyà" "Angostrina i Vilanova de les Escaldes" "Ansinhan" ...
#>  $ corregit      : logi  TRUE TRUE TRUE FALSE TRUE FALSE ...
#>  $ atles_CatNord : logi  TRUE TRUE TRUE FALSE TRUE TRUE ...
#>  $ ign_geoportail: logi  TRUE TRUE TRUE FALSE TRUE TRUE ...
#>  $ ign_mapes     : chr  "parcialment corregit" "corregit" "corregit" "no corregit" ...
#>  $ catala_pdf    : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...