Tesaurus de comarques de Catalunya Nord

Correspondències entre els municipis de les revisions de topònims dels cadastres feta per Joan Becat i OpenStreetMap.

Usage

tesaurus_comarques

Format

Un «data frame» amb 8 observacions de les següents 4 variables.

becat_nom

Nom de la comarca segons la revisió de topònims de Joan Becat.

osm_name

Nom de la comarca a OpenStreetMap.

osm_type

Tipus de l'objecte d'OSM que representa la comarca.

osm_id

Identificador de l'objecte d'OSM que representa la comarca.

Details

Source

data-raw/tesaurus_comarques.R

See also

Examples

str(tesaurus_comarques)
#> 'data.frame':	8 obs. of  4 variables:
#>  $ becat_nom: chr  "Capcir" "Cerdanya" "Conflent" "Fenolhedés" ...
#>  $ osm_name : chr  "Capcir" "Alta Cerdanya" "Conflent" "Fenolledès" ...
#>  $ osm_type : chr  "relation" "relation" "relation" "relation" ...
#>  $ osm_id   : chr  "13433227" "13433224" "13433226" "16780160" ...