Metadades de les revisions dels cadastres

Informació sistematitzada extreta de intro_cadastre.

Usage

meta_cadastre

Format

Un «data frame» amb 223 observacions de les següents 7 variables.

municipi

Nom del municipi.

corregit

Vector logic que indica si el cadastre ha estat corregit.

atles_CatNord

Vector logic que indica si els topònims apareixen a l’Atles toponímic de Catalunya Nord, ed. Terra Nostra, Prada, 2015, 2 volums, 970p.

ign_geoportail

Vector logic que indica si els topònims han estat corregits al GEOPORTAIL de l'IGN.

ign_mapes

Vector de tipus caràcter indicant si l'IGN a corregit els mapes o no, o si només parcialment.

catala_pdf

Vector logic que indica si el fitxer pdf és en català (TRUE) o en francès (FALSE).

dues_taules

Vector logic que indica si el municipi està dividit en dues taules.

Details

Els municipis amb dues_taules == TRUE, tenen dues taules a la llista becat_cadastre amb el nom del municipi seguit de _CADASTRE, _PROPOSAT, _CORREGIT o _NO CORREGIT.

Source

Joan Becat https://joanbecat.cat. Extracció de les dades a data-raw/metadades_becat_cadastre.R.

See also

Examples

str(meta_cadastre)
#> 'data.frame':	223 obs. of  7 variables:
#>  $ municipi      : chr  "Aiguatèbia i Talau" "Alenyà" "Angostrina i Vilanova de les Escaldes" "Ansinhan" ...
#>  $ corregit      : logi  TRUE TRUE TRUE FALSE TRUE FALSE ...
#>  $ atles_CatNord : logi  TRUE TRUE TRUE FALSE TRUE TRUE ...
#>  $ ign_geoportail: logi  TRUE TRUE TRUE FALSE TRUE TRUE ...
#>  $ ign_mapes     : chr  "parcialment corregit" "corregit" "corregit" "no corregit" ...
#>  $ catala_pdf    : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ dues_taules   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...