Tesaurus de municipis de Catalunya Nord

Correspondències entre els municipis de les revisions de topònims dels cadastres feta per Joan Becat i OpenStreetMap.

Usage

tesaurus_municipis

Format

Un «data frame» amb 223 observacions de les següents 7 variables.

becat_nom

Nom del municipi segons la revisió de topònims de Joan Becat.

becat_comarca

Nom de la comarca segons la revisió de topònims de Joan Becat.

osm_name:ca

Nom del municipi en català a OpenStreetMap.

osm_name

Nom del municipi a OpenStreetMap.

osm_comarca

a character vector

osm_type

Tipus de l'objecte d'OSM que representa el municipi.

osm_id

Identificador de l'objecte d'OSM que representa el municipi.

Details

Source

data-raw/tesaurus_municipis.R

See also

Examples

str(tesaurus_municipis)
#> 'data.frame':	223 obs. of  7 variables:
#>  $ becat_nom    : chr  "els Angles" "Font-rabiosa" "Formiguera" "Matamala" ...
#>  $ becat_comarca: chr  "Capcir" "Capcir" "Capcir" "Capcir" ...
#>  $ osm_name:ca  : chr  "els Angles" "Font-rabiosa" "Formiguera" "Matamala" ...
#>  $ osm_name     : chr  "Les Angles" "Fontrabiouse" "Formiguères" "Matemale" ...
#>  $ osm_comarca  : chr  "Capcir" "Capcir" "Capcir" "Capcir" ...
#>  $ osm_type     : chr  "relation" "relation" "relation" "relation" ...
#>  $ osm_id       : chr  "53365" "2814982" "2815058" "1069580" ...

discrepancies <- tesaurus_municipis[
  tesaurus_municipis$becat_nom != tesaurus_municipis$`osm_name:ca` &
    tesaurus_municipis$osm_comarca != "Fenolledès",
  c("becat_nom", "osm_name:ca", "osm_name", "osm_comarca")
]
discrepancies
#>                  becat_nom                osm_name:ca
#> 5              Puigvaledor                Puigbalador
#> 6                      Ral                       Real
#> 15   Font-romeu Odelló Vià   Font-romeu, Odelló i Vià
#> 32                 Castell          Castell de Vernet
#> 37         Conat i Vellans                      Conat
#> 39               els Masos                  Los Masos
#> 43              Eus i Coma                        Eus
#> 63                      Pi             Pi de Conflent
#> 70                   Sansà                      Censà
#> 72          Saorra i Toren                     Saorra
#> 73          Sautó i Fetges                      Sautó
#> 75       Soanyes i Marians                    Soanyes
#> 111 Argelers de la Marenda                   Argelers
#> 112                  Bages          Bages de Rosselló
#> 116                    Baó                        Bao
#> 117                 Bonpàs                     Bompàs
#> 125                  Canet          Canet de Rosselló
#> 130                Cervera      Cervera de la Marenda
#> 132                Corbera        Corbera de Rosselló
#> 133      Corbera la Cabana       la Cabana de Corbera
#> 145               l'Albera                   L'Albera
#> 147    la Roca de l'Albera           la Roca d'Albera
#> 148     la Torre del Bisbe            la Torre d'Elna
#> 154             Montesquiu        Montesquiu d'Albera
#> 168       Pontellà i Nyils                   Pontellà
#> 175            Sant Cebrià    Sant Cebrià de Rosselló
#> 180           Sant Hipòlit Sant Hipòlit de la Salanca
#> 184            Sant Nazari    Sant Nazari de Rosselló
#> 185           Santa Coloma       Santa Coloma de Tuïr
#> 193              Torrelles    Torrelles de la Salanca
#> 201        Vilanova de Raó            Vilanova de Raò
#> 203      Prunet i Bellpuig           Prunet i Belpuig
#> 207                Custoja                    Costoja
#> 215         Prats de Molló Prats de Molló i la Presta
#> 217 Sant Joan Pla de Corts    Sant Joan de Pladecorts
#>                             osm_name   osm_comarca
#> 5                         Puyvalador        Capcir
#> 6                               Réal        Capcir
#> 15            Font-Romeu-Odeillo-Via Alta Cerdanya
#> 32                           Casteil      Conflent
#> 37                             Conat      Conflent
#> 39                         Los Masos      Conflent
#> 43                               Eus      Conflent
#> 63                                Py      Conflent
#> 70                             Sansa      Conflent
#> 72                           Sahorre      Conflent
#> 73                             Sauto      Conflent
#> 75                          Souanyas      Conflent
#> 111                  Argelès-sur-Mer      Rosselló
#> 112                            Bages      Rosselló
#> 116                             Baho      Rosselló
#> 117                           Bompas      Rosselló
#> 125              Canet-en-Roussillon      Rosselló
#> 130                          Cerbère      Rosselló
#> 132                          Corbère      Rosselló
#> 133              Corbère-les-Cabanes      Rosselló
#> 145                         L'Albère      Rosselló
#> 147              Laroque-des-Albères      Rosselló
#> 148                  Latour-Bas-Elne      Rosselló
#> 154          Montesquieu-des-Albères      Rosselló
#> 168                        Ponteilla      Rosselló
#> 175                    Saint-Cyprien      Rosselló
#> 180                  Saint-Hippolyte      Rosselló
#> 184                    Saint-Nazaire      Rosselló
#> 185 Sainte-Colombe-de-la-Commanderie      Rosselló
#> 193                       Torreilles      Rosselló
#> 201            Villeneuve-de-la-Raho      Rosselló
#> 203                Prunet-et-Belpuig      Rosselló
#> 207                       Coustouges     Vallespir
#> 215         Prats-de-Mollo-la-Preste     Vallespir
#> 217          Saint-Jean-Pla-de-Corts     Vallespir