Tesaurus de municipis de Catalunya Nord

Correspondències entre les comarques de les revisions de topònims dels cadastres feta per Joan Becat i OpenstreetMap.

Ús

tesaurus_municipis

Format

Un «data frame» amb 223 observacions de les següents 6 variables.

becat_nom

Nom del municipi segons la revisió de topònims de Joan Becat.

becat_comarca

Nom de la comarca segons la revisió de topònims de Joan Becat.

osm_name

Nom del municipi a OpenstreetMap.

osm_comarca

a character vector

osm_type

Tipus de l'objecte d'OSM que representa el municipi.

osm_id

Identificador de l'objecte d'OSM que representa el municipi.

Detalls

Font

data-raw/tesaurus_municipis.R

Exemples

str(tesaurus_municipis)
#> 'data.frame':	223 obs. of  6 variables:
#>  $ becat_nom    : chr  "Aiguatèbia i Talau" "Alenyà" "Angostrina i Vilanova de les Escaldes" "Ansinhan" ...
#>  $ becat_comarca: chr  "Conflent" "Rosselló" "Cerdanya" "Fenolhedés" ...
#>  $ osm_name     : chr  "Aiguatèbia i Talau" "Alenyà" "Angostrina i Vilanova de les Escaldes" "Ansignan" ...
#>  $ osm_comarca  : chr  "Conflent" "Rosselló" "Alta Cerdanya" "Fenolledès" ...
#>  $ osm_type     : chr  "relation" "relation" "relation" "relation" ...
#>  $ osm_id       : chr  "2912344" "271664" "2814313" "21627" ...